Relocating activity classes to activity package
