+++ src/main/java/com/ninjaone/dundie_awards/activity/ActivityRepository.java
@@ -1,6 +1,5 @@
-package com.ninjaone.dundie_awards.repository;
+package com.ninjaone.dundie_awards.activity;
 
-import com.ninjaone.dundie_awards.model.Activity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
