+++ src/main/java/com/ninjaone/dundie_awards/controller/IndexController.java
@@ -1,9 +1,9 @@
 package com.ninjaone.dundie_awards.controller;
 
 import com.ninjaone.dundie_awards.AwardsCache;
-import com.ninjaone.dundie_awards.MessageBroker;
+import com.ninjaone.dundie_awards.activity.MessageBroker;
 import com.ninjaone.dundie_awards.employee.EmployeeRepository;
-import com.ninjaone.dundie_awards.repository.ActivityRepository;
+import com.ninjaone.dundie_awards.activity.ActivityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
