+++ src/main/java/com/ninjaone/dundie_awards/employee/EmployeeController.java
@@ -1,8 +1,8 @@
 package com.ninjaone.dundie_awards.employee;
 
 import com.ninjaone.dundie_awards.AwardsCache;
-import com.ninjaone.dundie_awards.MessageBroker;
-import com.ninjaone.dundie_awards.repository.ActivityRepository;
+import com.ninjaone.dundie_awards.activity.MessageBroker;
+import com.ninjaone.dundie_awards.activity.ActivityRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
